Move unit tests to its own folder
