+++ tests/Unit/CacheComponentTest.php
