+++ tests/Unit/EventTriggerTest.php
